將 Food 模型的 carbGrams 欄位型別由 Integer 改為 Double，以支援小數值

+++ src/main/java/com/example/foodhistory/model/Food.java
@@ -18,7 +18,7 @@ public class Food {
     private String servingSize;
     
     @Column(name = "carb_grams")
-    private Integer carbGrams;
+    private Double carbGrams;
     
     @Column
     private Double quantity;
